Struct sagemaker_helloworld::Client

pub(crate) struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> {
    handle: Arc<Handle<C, M, R>>,
}

An ergonomic service client for SageMaker.

This client allows ergonomic access to a SageMaker-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Fields
handle: Arc<Handle<C, M, R>>
Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Client<C, M, R>

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_association(&self) -> AddAssociation<C, M, R>

Constructs a fluent builder for the AddAssociation operation.

See AddAssociation for more information about the operation and its arguments.

pub fn add_tags(&self) -> AddTags<C, M, R>

Constructs a fluent builder for the AddTags operation.

See AddTags for more information about the operation and its arguments.

pub fn associate_trial_component(&self) -> AssociateTrialComponent<C, M, R>

Constructs a fluent builder for the AssociateTrialComponent operation.

See AssociateTrialComponent for more information about the operation and its arguments.

pub fn batch_describe_model_package(&self) -> BatchDescribeModelPackage<C, M, R>

Constructs a fluent builder for the BatchDescribeModelPackage operation.

See BatchDescribeModelPackage for more information about the operation and its arguments.

pub fn create_action(&self) -> CreateAction<C, M, R>

Constructs a fluent builder for the CreateAction operation.

See CreateAction for more information about the operation and its arguments.

pub fn create_algorithm(&self) -> CreateAlgorithm<C, M, R>

Constructs a fluent builder for the CreateAlgorithm operation.

See CreateAlgorithm for more information about the operation and its arguments.

pub fn create_app(&self) -> CreateApp<C, M, R>

Constructs a fluent builder for the CreateApp operation.

See CreateApp for more information about the operation and its arguments.

pub fn create_app_image_config(&self) -> CreateAppImageConfig<C, M, R>

Constructs a fluent builder for the CreateAppImageConfig operation.

See CreateAppImageConfig for more information about the operation and its arguments.

pub fn create_artifact(&self) -> CreateArtifact<C, M, R>

Constructs a fluent builder for the CreateArtifact operation.

See CreateArtifact for more information about the operation and its arguments.

pub fn create_auto_ml_job(&self) -> CreateAutoMLJob<C, M, R>

Constructs a fluent builder for the CreateAutoMLJob operation.

See CreateAutoMLJob for more information about the operation and its arguments.

pub fn create_code_repository(&self) -> CreateCodeRepository<C, M, R>

Constructs a fluent builder for the CreateCodeRepository operation.

See CreateCodeRepository for more information about the operation and its arguments.

pub fn create_compilation_job(&self) -> CreateCompilationJob<C, M, R>

Constructs a fluent builder for the CreateCompilationJob operation.

See CreateCompilationJob for more information about the operation and its arguments.

pub fn create_context(&self) -> CreateContext<C, M, R>

Constructs a fluent builder for the CreateContext operation.

See CreateContext for more information about the operation and its arguments.

pub fn create_data_quality_job_definition(
&self
) -> CreateDataQualityJobDefinition<C, M, R>

Constructs a fluent builder for the CreateDataQualityJobDefinition operation.

See CreateDataQualityJobDefinition for more information about the operation and its arguments.

pub fn create_device_fleet(&self) -> CreateDeviceFleet<C, M, R>

Constructs a fluent builder for the CreateDeviceFleet operation.

See CreateDeviceFleet for more information about the operation and its arguments.

pub fn create_domain(&self) -> CreateDomain<C, M, R>

Constructs a fluent builder for the CreateDomain operation.

See CreateDomain for more information about the operation and its arguments.

pub fn create_edge_packaging_job(&self) -> CreateEdgePackagingJob<C, M, R>

Constructs a fluent builder for the CreateEdgePackagingJob operation.

See CreateEdgePackagingJob for more information about the operation and its arguments.

pub fn create_endpoint(&self) -> CreateEndpoint<C, M, R>

Constructs a fluent builder for the CreateEndpoint operation.

See CreateEndpoint for more information about the operation and its arguments.

pub fn create_endpoint_config(&self) -> CreateEndpointConfig<C, M, R>

Constructs a fluent builder for the CreateEndpointConfig operation.

See CreateEndpointConfig for more information about the operation and its arguments.

pub fn create_experiment(&self) -> CreateExperiment<C, M, R>

Constructs a fluent builder for the CreateExperiment operation.

See CreateExperiment for more information about the operation and its arguments.

pub fn create_feature_group(&self) -> CreateFeatureGroup<C, M, R>

Constructs a fluent builder for the CreateFeatureGroup operation.

See CreateFeatureGroup for more information about the operation and its arguments.

pub fn create_flow_definition(&self) -> CreateFlowDefinition<C, M, R>

Constructs a fluent builder for the CreateFlowDefinition operation.

See CreateFlowDefinition for more information about the operation and its arguments.

pub fn create_human_task_ui(&self) -> CreateHumanTaskUi<C, M, R>

Constructs a fluent builder for the CreateHumanTaskUi operation.

See CreateHumanTaskUi for more information about the operation and its arguments.

pub fn create_hyper_parameter_tuning_job(
&self
) -> CreateHyperParameterTuningJob<C, M, R>

Constructs a fluent builder for the CreateHyperParameterTuningJob operation.

See CreateHyperParameterTuningJob for more information about the operation and its arguments.

pub fn create_image(&self) -> CreateImage<C, M, R>

Constructs a fluent builder for the CreateImage operation.

See CreateImage for more information about the operation and its arguments.

pub fn create_image_version(&self) -> CreateImageVersion<C, M, R>

Constructs a fluent builder for the CreateImageVersion operation.

See CreateImageVersion for more information about the operation and its arguments.

pub fn create_labeling_job(&self) -> CreateLabelingJob<C, M, R>

Constructs a fluent builder for the CreateLabelingJob operation.

See CreateLabelingJob for more information about the operation and its arguments.

pub fn create_model(&self) -> CreateModel<C, M, R>

Constructs a fluent builder for the CreateModel operation.

See CreateModel for more information about the operation and its arguments.

pub fn create_model_bias_job_definition(
&self
) -> CreateModelBiasJobDefinition<C, M, R>

Constructs a fluent builder for the CreateModelBiasJobDefinition operation.

See CreateModelBiasJobDefinition for more information about the operation and its arguments.

pub fn create_model_explainability_job_definition(
&self
) -> CreateModelExplainabilityJobDefinition<C, M, R>

Constructs a fluent builder for the CreateModelExplainabilityJobDefinition operation.

See CreateModelExplainabilityJobDefinition for more information about the operation and its arguments.

pub fn create_model_package(&self) -> CreateModelPackage<C, M, R>

Constructs a fluent builder for the CreateModelPackage operation.

See CreateModelPackage for more information about the operation and its arguments.

pub fn create_model_package_group(&self) -> CreateModelPackageGroup<C, M, R>

Constructs a fluent builder for the CreateModelPackageGroup operation.

See CreateModelPackageGroup for more information about the operation and its arguments.

pub fn create_model_quality_job_definition(
&self
) -> CreateModelQualityJobDefinition<C, M, R>

Constructs a fluent builder for the CreateModelQualityJobDefinition operation.

See CreateModelQualityJobDefinition for more information about the operation and its arguments.

pub fn create_monitoring_schedule(&self) -> CreateMonitoringSchedule<C, M, R>

Constructs a fluent builder for the CreateMonitoringSchedule operation.

See CreateMonitoringSchedule for more information about the operation and its arguments.

pub fn create_notebook_instance(&self) -> CreateNotebookInstance<C, M, R>

Constructs a fluent builder for the CreateNotebookInstance operation.

See CreateNotebookInstance for more information about the operation and its arguments.

pub fn create_notebook_instance_lifecycle_config(
&self
) -> CreateNotebookInstanceLifecycleConfig<C, M, R>

Constructs a fluent builder for the CreateNotebookInstanceLifecycleConfig operation.

See CreateNotebookInstanceLifecycleConfig for more information about the operation and its arguments.

pub fn create_pipeline(&self) -> CreatePipeline<C, M, R>

Constructs a fluent builder for the CreatePipeline operation.

See CreatePipeline for more information about the operation and its arguments.

pub fn create_presigned_domain_url(&self) -> CreatePresignedDomainUrl<C, M, R>

Constructs a fluent builder for the CreatePresignedDomainUrl operation.

See CreatePresignedDomainUrl for more information about the operation and its arguments.

pub fn create_presigned_notebook_instance_url(
&self
) -> CreatePresignedNotebookInstanceUrl<C, M, R>

Constructs a fluent builder for the CreatePresignedNotebookInstanceUrl operation.

See CreatePresignedNotebookInstanceUrl for more information about the operation and its arguments.

pub fn create_processing_job(&self) -> CreateProcessingJob<C, M, R>

Constructs a fluent builder for the CreateProcessingJob operation.

See CreateProcessingJob for more information about the operation and its arguments.

pub fn create_project(&self) -> CreateProject<C, M, R>

Constructs a fluent builder for the CreateProject operation.

See CreateProject for more information about the operation and its arguments.

pub fn create_studio_lifecycle_config(
&self
) -> CreateStudioLifecycleConfig<C, M, R>

Constructs a fluent builder for the CreateStudioLifecycleConfig operation.

See CreateStudioLifecycleConfig for more information about the operation and its arguments.

pub fn create_training_job(&self) -> CreateTrainingJob<C, M, R>

Constructs a fluent builder for the CreateTrainingJob operation.

See CreateTrainingJob for more information about the operation and its arguments.

pub fn create_transform_job(&self) -> CreateTransformJob<C, M, R>

Constructs a fluent builder for the CreateTransformJob operation.

See CreateTransformJob for more information about the operation and its arguments.

pub fn create_trial(&self) -> CreateTrial<C, M, R>

Constructs a fluent builder for the CreateTrial operation.

See CreateTrial for more information about the operation and its arguments.

pub fn create_trial_component(&self) -> CreateTrialComponent<C, M, R>

Constructs a fluent builder for the CreateTrialComponent operation.

See CreateTrialComponent for more information about the operation and its arguments.

pub fn create_user_profile(&self) -> CreateUserProfile<C, M, R>

Constructs a fluent builder for the CreateUserProfile operation.

See CreateUserProfile for more information about the operation and its arguments.

pub fn create_workforce(&self) -> CreateWorkforce<C, M, R>

Constructs a fluent builder for the CreateWorkforce operation.

See CreateWorkforce for more information about the operation and its arguments.

pub fn create_workteam(&self) -> CreateWorkteam<C, M, R>

Constructs a fluent builder for the CreateWorkteam operation.

See CreateWorkteam for more information about the operation and its arguments.

pub fn delete_action(&self) -> DeleteAction<C, M, R>

Constructs a fluent builder for the DeleteAction operation.

See DeleteAction for more information about the operation and its arguments.

pub fn delete_algorithm(&self) -> DeleteAlgorithm<C, M, R>

Constructs a fluent builder for the DeleteAlgorithm operation.

See DeleteAlgorithm for more information about the operation and its arguments.

pub fn delete_app(&self) -> DeleteApp<C, M, R>

Constructs a fluent builder for the DeleteApp operation.

See DeleteApp for more information about the operation and its arguments.

pub fn delete_app_image_config(&self) -> DeleteAppImageConfig<C, M, R>

Constructs a fluent builder for the DeleteAppImageConfig operation.

See DeleteAppImageConfig for more information about the operation and its arguments.

pub fn delete_artifact(&self) -> DeleteArtifact<C, M, R>

Constructs a fluent builder for the DeleteArtifact operation.

See DeleteArtifact for more information about the operation and its arguments.

pub fn delete_association(&self) -> DeleteAssociation<C, M, R>

Constructs a fluent builder for the DeleteAssociation operation.

See DeleteAssociation for more information about the operation and its arguments.

pub fn delete_code_repository(&self) -> DeleteCodeRepository<C, M, R>

Constructs a fluent builder for the DeleteCodeRepository operation.

See DeleteCodeRepository for more information about the operation and its arguments.

pub fn delete_context(&self) -> DeleteContext<C, M, R>

Constructs a fluent builder for the DeleteContext operation.

See DeleteContext for more information about the operation and its arguments.

pub fn delete_data_quality_job_definition(
&self
) -> DeleteDataQualityJobDefinition<C, M, R>

Constructs a fluent builder for the DeleteDataQualityJobDefinition operation.

See DeleteDataQualityJobDefinition for more information about the operation and its arguments.

pub fn delete_device_fleet(&self) -> DeleteDeviceFleet<C, M, R>

Constructs a fluent builder for the DeleteDeviceFleet operation.

See DeleteDeviceFleet for more information about the operation and its arguments.

pub fn delete_domain(&self) -> DeleteDomain<C, M, R>

Constructs a fluent builder for the DeleteDomain operation.

See DeleteDomain for more information about the operation and its arguments.

pub fn delete_endpoint(&self) -> DeleteEndpoint<C, M, R>

Constructs a fluent builder for the DeleteEndpoint operation.

See DeleteEndpoint for more information about the operation and its arguments.

pub fn delete_endpoint_config(&self) -> DeleteEndpointConfig<C, M, R>

Constructs a fluent builder for the DeleteEndpointConfig operation.

See DeleteEndpointConfig for more information about the operation and its arguments.

pub fn delete_experiment(&self) -> DeleteExperiment<C, M, R>

Constructs a fluent builder for the DeleteExperiment operation.

See DeleteExperiment for more information about the operation and its arguments.

pub fn delete_feature_group(&self) -> DeleteFeatureGroup<C, M, R>

Constructs a fluent builder for the DeleteFeatureGroup operation.

See DeleteFeatureGroup for more information about the operation and its arguments.

pub fn delete_flow_definition(&self) -> DeleteFlowDefinition<C, M, R>

Constructs a fluent builder for the DeleteFlowDefinition operation.

See DeleteFlowDefinition for more information about the operation and its arguments.

pub fn delete_human_task_ui(&self) -> DeleteHumanTaskUi<C, M, R>

Constructs a fluent builder for the DeleteHumanTaskUi operation.

See DeleteHumanTaskUi for more information about the operation and its arguments.

pub fn delete_image(&self) -> DeleteImage<C, M, R>

Constructs a fluent builder for the DeleteImage operation.

See DeleteImage for more information about the operation and its arguments.

pub fn delete_image_version(&self) -> DeleteImageVersion<C, M, R>

Constructs a fluent builder for the DeleteImageVersion operation.

See DeleteImageVersion for more information about the operation and its arguments.

pub fn delete_model(&self) -> DeleteModel<C, M, R>

Constructs a fluent builder for the DeleteModel operation.

See DeleteModel for more information about the operation and its arguments.

pub fn delete_model_bias_job_definition(
&self
) -> DeleteModelBiasJobDefinition<C, M, R>

Constructs a fluent builder for the DeleteModelBiasJobDefinition operation.

See DeleteModelBiasJobDefinition for more information about the operation and its arguments.

pub fn delete_model_explainability_job_definition(
&self
) -> DeleteModelExplainabilityJobDefinition<C, M, R>

Constructs a fluent builder for the DeleteModelExplainabilityJobDefinition operation.

See DeleteModelExplainabilityJobDefinition for more information about the operation and its arguments.

pub fn delete_model_package(&self) -> DeleteModelPackage<C, M, R>

Constructs a fluent builder for the DeleteModelPackage operation.

See DeleteModelPackage for more information about the operation and its arguments.

pub fn delete_model_package_group(&self) -> DeleteModelPackageGroup<C, M, R>

Constructs a fluent builder for the DeleteModelPackageGroup operation.

See DeleteModelPackageGroup for more information about the operation and its arguments.

pub fn delete_model_package_group_policy(
&self
) -> DeleteModelPackageGroupPolicy<C, M, R>

Constructs a fluent builder for the DeleteModelPackageGroupPolicy operation.

See DeleteModelPackageGroupPolicy for more information about the operation and its arguments.

pub fn delete_model_quality_job_definition(
&self
) -> DeleteModelQualityJobDefinition<C, M, R>

Constructs a fluent builder for the DeleteModelQualityJobDefinition operation.

See DeleteModelQualityJobDefinition for more information about the operation and its arguments.

pub fn delete_monitoring_schedule(&self) -> DeleteMonitoringSchedule<C, M, R>

Constructs a fluent builder for the DeleteMonitoringSchedule operation.

See DeleteMonitoringSchedule for more information about the operation and its arguments.

pub fn delete_notebook_instance(&self) -> DeleteNotebookInstance<C, M, R>

Constructs a fluent builder for the DeleteNotebookInstance operation.

See DeleteNotebookInstance for more information about the operation and its arguments.

pub fn delete_notebook_instance_lifecycle_config(
&self
) -> DeleteNotebookInstanceLifecycleConfig<C, M, R>

Constructs a fluent builder for the DeleteNotebookInstanceLifecycleConfig operation.

See DeleteNotebookInstanceLifecycleConfig for more information about the operation and its arguments.

pub fn delete_pipeline(&self) -> DeletePipeline<C, M, R>

Constructs a fluent builder for the DeletePipeline operation.

See DeletePipeline for more information about the operation and its arguments.

pub fn delete_project(&self) -> DeleteProject<C, M, R>

Constructs a fluent builder for the DeleteProject operation.

See DeleteProject for more information about the operation and its arguments.

pub fn delete_studio_lifecycle_config(
&self
) -> DeleteStudioLifecycleConfig<C, M, R>

Constructs a fluent builder for the DeleteStudioLifecycleConfig operation.

See DeleteStudioLifecycleConfig for more information about the operation and its arguments.

pub fn delete_tags(&self) -> DeleteTags<C, M, R>

Constructs a fluent builder for the DeleteTags operation.

See DeleteTags for more information about the operation and its arguments.

pub fn delete_trial(&self) -> DeleteTrial<C, M, R>

Constructs a fluent builder for the DeleteTrial operation.

See DeleteTrial for more information about the operation and its arguments.

pub fn delete_trial_component(&self) -> DeleteTrialComponent<C, M, R>

Constructs a fluent builder for the DeleteTrialComponent operation.

See DeleteTrialComponent for more information about the operation and its arguments.

pub fn delete_user_profile(&self) -> DeleteUserProfile<C, M, R>

Constructs a fluent builder for the DeleteUserProfile operation.

See DeleteUserProfile for more information about the operation and its arguments.

pub fn delete_workforce(&self) -> DeleteWorkforce<C, M, R>

Constructs a fluent builder for the DeleteWorkforce operation.

See DeleteWorkforce for more information about the operation and its arguments.

pub fn delete_workteam(&self) -> DeleteWorkteam<C, M, R>

Constructs a fluent builder for the DeleteWorkteam operation.

See DeleteWorkteam for more information about the operation and its arguments.

pub fn deregister_devices(&self) -> DeregisterDevices<C, M, R>

Constructs a fluent builder for the DeregisterDevices operation.

See DeregisterDevices for more information about the operation and its arguments.

pub fn describe_action(&self) -> DescribeAction<C, M, R>

Constructs a fluent builder for the DescribeAction operation.

See DescribeAction for more information about the operation and its arguments.

pub fn describe_algorithm(&self) -> DescribeAlgorithm<C, M, R>

Constructs a fluent builder for the DescribeAlgorithm operation.

See DescribeAlgorithm for more information about the operation and its arguments.

pub fn describe_app(&self) -> DescribeApp<C, M, R>

Constructs a fluent builder for the DescribeApp operation.

See DescribeApp for more information about the operation and its arguments.

pub fn describe_app_image_config(&self) -> DescribeAppImageConfig<C, M, R>

Constructs a fluent builder for the DescribeAppImageConfig operation.

See DescribeAppImageConfig for more information about the operation and its arguments.

pub fn describe_artifact(&self) -> DescribeArtifact<C, M, R>

Constructs a fluent builder for the DescribeArtifact operation.

See DescribeArtifact for more information about the operation and its arguments.

pub fn describe_auto_ml_job(&self) -> DescribeAutoMLJob<C, M, R>

Constructs a fluent builder for the DescribeAutoMLJob operation.

See DescribeAutoMLJob for more information about the operation and its arguments.

pub fn describe_code_repository(&self) -> DescribeCodeRepository<C, M, R>

Constructs a fluent builder for the DescribeCodeRepository operation.

See DescribeCodeRepository for more information about the operation and its arguments.

pub fn describe_compilation_job(&self) -> DescribeCompilationJob<C, M, R>

Constructs a fluent builder for the DescribeCompilationJob operation.

See DescribeCompilationJob for more information about the operation and its arguments.

pub fn describe_context(&self) -> DescribeContext<C, M, R>

Constructs a fluent builder for the DescribeContext operation.

See DescribeContext for more information about the operation and its arguments.

pub fn describe_data_quality_job_definition(
&self
) -> DescribeDataQualityJobDefinition<C, M, R>

Constructs a fluent builder for the DescribeDataQualityJobDefinition operation.

See DescribeDataQualityJobDefinition for more information about the operation and its arguments.

pub fn describe_device(&self) -> DescribeDevice<C, M, R>

Constructs a fluent builder for the DescribeDevice operation.

See DescribeDevice for more information about the operation and its arguments.

pub fn describe_device_fleet(&self) -> DescribeDeviceFleet<C, M, R>

Constructs a fluent builder for the DescribeDeviceFleet operation.

See DescribeDeviceFleet for more information about the operation and its arguments.

pub fn describe_domain(&self) -> DescribeDomain<C, M, R>

Constructs a fluent builder for the DescribeDomain operation.

See DescribeDomain for more information about the operation and its arguments.

pub fn describe_edge_packaging_job(&self) -> DescribeEdgePackagingJob<C, M, R>

Constructs a fluent builder for the DescribeEdgePackagingJob operation.

See DescribeEdgePackagingJob for more information about the operation and its arguments.

pub fn describe_endpoint(&self) -> DescribeEndpoint<C, M, R>

Constructs a fluent builder for the DescribeEndpoint operation.

See DescribeEndpoint for more information about the operation and its arguments.

pub fn describe_endpoint_config(&self) -> DescribeEndpointConfig<C, M, R>

Constructs a fluent builder for the DescribeEndpointConfig operation.

See DescribeEndpointConfig for more information about the operation and its arguments.

pub fn describe_experiment(&self) -> DescribeExperiment<C, M, R>

Constructs a fluent builder for the DescribeExperiment operation.

See DescribeExperiment for more information about the operation and its arguments.

pub fn describe_feature_group(&self) -> DescribeFeatureGroup<C, M, R>

Constructs a fluent builder for the DescribeFeatureGroup operation.

See DescribeFeatureGroup for more information about the operation and its arguments.

pub fn describe_flow_definition(&self) -> DescribeFlowDefinition<C, M, R>

Constructs a fluent builder for the DescribeFlowDefinition operation.

See DescribeFlowDefinition for more information about the operation and its arguments.

pub fn describe_human_task_ui(&self) -> DescribeHumanTaskUi<C, M, R>

Constructs a fluent builder for the DescribeHumanTaskUi operation.

See DescribeHumanTaskUi for more information about the operation and its arguments.

pub fn describe_hyper_parameter_tuning_job(
&self
) -> DescribeHyperParameterTuningJob<C, M, R>

Constructs a fluent builder for the DescribeHyperParameterTuningJob operation.

See DescribeHyperParameterTuningJob for more information about the operation and its arguments.

pub fn describe_image(&self) -> DescribeImage<C, M, R>

Constructs a fluent builder for the DescribeImage operation.

See DescribeImage for more information about the operation and its arguments.

pub fn describe_image_version(&self) -> DescribeImageVersion<C, M, R>

Constructs a fluent builder for the DescribeImageVersion operation.

See DescribeImageVersion for more information about the operation and its arguments.

pub fn describe_labeling_job(&self) -> DescribeLabelingJob<C, M, R>

Constructs a fluent builder for the DescribeLabelingJob operation.

See DescribeLabelingJob for more information about the operation and its arguments.

pub fn describe_model(&self) -> DescribeModel<C, M, R>

Constructs a fluent builder for the DescribeModel operation.

See DescribeModel for more information about the operation and its arguments.

pub fn describe_model_bias_job_definition(
&self
) -> DescribeModelBiasJobDefinition<C, M, R>

Constructs a fluent builder for the DescribeModelBiasJobDefinition operation.

See DescribeModelBiasJobDefinition for more information about the operation and its arguments.

pub fn describe_model_explainability_job_definition(
&self
) -> DescribeModelExplainabilityJobDefinition<C, M, R>

Constructs a fluent builder for the DescribeModelExplainabilityJobDefinition operation.

See DescribeModelExplainabilityJobDefinition for more information about the operation and its arguments.

pub fn describe_model_package(&self) -> DescribeModelPackage<C, M, R>

Constructs a fluent builder for the DescribeModelPackage operation.

See DescribeModelPackage for more information about the operation and its arguments.

pub fn describe_model_package_group(&self) -> DescribeModelPackageGroup<C, M, R>

Constructs a fluent builder for the DescribeModelPackageGroup operation.

See DescribeModelPackageGroup for more information about the operation and its arguments.

pub fn describe_model_quality_job_definition(
&self
) -> DescribeModelQualityJobDefinition<C, M, R>

Constructs a fluent builder for the DescribeModelQualityJobDefinition operation.

See DescribeModelQualityJobDefinition for more information about the operation and its arguments.

pub fn describe_monitoring_schedule(
&self
) -> DescribeMonitoringSchedule<C, M, R>

Constructs a fluent builder for the DescribeMonitoringSchedule operation.

See DescribeMonitoringSchedule for more information about the operation and its arguments.

pub fn describe_notebook_instance(&self) -> DescribeNotebookInstance<C, M, R>

Constructs a fluent builder for the DescribeNotebookInstance operation.

See DescribeNotebookInstance for more information about the operation and its arguments.

pub fn describe_notebook_instance_lifecycle_config(
&self
) -> DescribeNotebookInstanceLifecycleConfig<C, M, R>

Constructs a fluent builder for the DescribeNotebookInstanceLifecycleConfig operation.

See DescribeNotebookInstanceLifecycleConfig for more information about the operation and its arguments.

pub fn describe_pipeline(&self) -> DescribePipeline<C, M, R>

Constructs a fluent builder for the DescribePipeline operation.

See DescribePipeline for more information about the operation and its arguments.

pub fn describe_pipeline_definition_for_execution(
&self
) -> DescribePipelineDefinitionForExecution<C, M, R>

Constructs a fluent builder for the DescribePipelineDefinitionForExecution operation.

See DescribePipelineDefinitionForExecution for more information about the operation and its arguments.

pub fn describe_pipeline_execution(&self) -> DescribePipelineExecution<C, M, R>

Constructs a fluent builder for the DescribePipelineExecution operation.

See DescribePipelineExecution for more information about the operation and its arguments.

pub fn describe_processing_job(&self) -> DescribeProcessingJob<C, M, R>

Constructs a fluent builder for the DescribeProcessingJob operation.

See DescribeProcessingJob for more information about the operation and its arguments.

pub fn describe_project(&self) -> DescribeProject<C, M, R>

Constructs a fluent builder for the DescribeProject operation.

See DescribeProject for more information about the operation and its arguments.

pub fn describe_studio_lifecycle_config(
&self
) -> DescribeStudioLifecycleConfig<C, M, R>

Constructs a fluent builder for the DescribeStudioLifecycleConfig operation.

See DescribeStudioLifecycleConfig for more information about the operation and its arguments.

pub fn describe_subscribed_workteam(
&self
) -> DescribeSubscribedWorkteam<C, M, R>

Constructs a fluent builder for the DescribeSubscribedWorkteam operation.

See DescribeSubscribedWorkteam for more information about the operation and its arguments.

pub fn describe_training_job(&self) -> DescribeTrainingJob<C, M, R>

Constructs a fluent builder for the DescribeTrainingJob operation.

See DescribeTrainingJob for more information about the operation and its arguments.

pub fn describe_transform_job(&self) -> DescribeTransformJob<C, M, R>

Constructs a fluent builder for the DescribeTransformJob operation.

See DescribeTransformJob for more information about the operation and its arguments.

pub fn describe_trial(&self) -> DescribeTrial<C, M, R>

Constructs a fluent builder for the DescribeTrial operation.

See DescribeTrial for more information about the operation and its arguments.

pub fn describe_trial_component(&self) -> DescribeTrialComponent<C, M, R>

Constructs a fluent builder for the DescribeTrialComponent operation.

See DescribeTrialComponent for more information about the operation and its arguments.

pub fn describe_user_profile(&self) -> DescribeUserProfile<C, M, R>

Constructs a fluent builder for the DescribeUserProfile operation.

See DescribeUserProfile for more information about the operation and its arguments.

pub fn describe_workforce(&self) -> DescribeWorkforce<C, M, R>

Constructs a fluent builder for the DescribeWorkforce operation.

See DescribeWorkforce for more information about the operation and its arguments.

pub fn describe_workteam(&self) -> DescribeWorkteam<C, M, R>

Constructs a fluent builder for the DescribeWorkteam operation.

See DescribeWorkteam for more information about the operation and its arguments.

pub fn disable_sagemaker_servicecatalog_portfolio(
&self
) -> DisableSagemakerServicecatalogPortfolio<C, M, R>

Constructs a fluent builder for the DisableSagemakerServicecatalogPortfolio operation.

See DisableSagemakerServicecatalogPortfolio for more information about the operation and its arguments.

pub fn disassociate_trial_component(
&self
) -> DisassociateTrialComponent<C, M, R>

Constructs a fluent builder for the DisassociateTrialComponent operation.

See DisassociateTrialComponent for more information about the operation and its arguments.

pub fn enable_sagemaker_servicecatalog_portfolio(
&self
) -> EnableSagemakerServicecatalogPortfolio<C, M, R>

Constructs a fluent builder for the EnableSagemakerServicecatalogPortfolio operation.

See EnableSagemakerServicecatalogPortfolio for more information about the operation and its arguments.

pub fn get_device_fleet_report(&self) -> GetDeviceFleetReport<C, M, R>

Constructs a fluent builder for the GetDeviceFleetReport operation.

See GetDeviceFleetReport for more information about the operation and its arguments.

pub fn get_model_package_group_policy(
&self
) -> GetModelPackageGroupPolicy<C, M, R>

Constructs a fluent builder for the GetModelPackageGroupPolicy operation.

See GetModelPackageGroupPolicy for more information about the operation and its arguments.

pub fn get_sagemaker_servicecatalog_portfolio_status(
&self
) -> GetSagemakerServicecatalogPortfolioStatus<C, M, R>

Constructs a fluent builder for the GetSagemakerServicecatalogPortfolioStatus operation.

See GetSagemakerServicecatalogPortfolioStatus for more information about the operation and its arguments.

pub fn get_search_suggestions(&self) -> GetSearchSuggestions<C, M, R>

Constructs a fluent builder for the GetSearchSuggestions operation.

See GetSearchSuggestions for more information about the operation and its arguments.

pub fn list_actions(&self) -> ListActions<C, M, R>

Constructs a fluent builder for the ListActions operation.

See ListActions for more information about the operation and its arguments.

pub fn list_algorithms(&self) -> ListAlgorithms<C, M, R>

Constructs a fluent builder for the ListAlgorithms operation.

See ListAlgorithms for more information about the operation and its arguments.

pub fn list_app_image_configs(&self) -> ListAppImageConfigs<C, M, R>

Constructs a fluent builder for the ListAppImageConfigs operation.

See ListAppImageConfigs for more information about the operation and its arguments.

pub fn list_apps(&self) -> ListApps<C, M, R>

Constructs a fluent builder for the ListApps operation.

See ListApps for more information about the operation and its arguments.

pub fn list_artifacts(&self) -> ListArtifacts<C, M, R>

Constructs a fluent builder for the ListArtifacts operation.

See ListArtifacts for more information about the operation and its arguments.

pub fn list_associations(&self) -> ListAssociations<C, M, R>

Constructs a fluent builder for the ListAssociations operation.

See ListAssociations for more information about the operation and its arguments.

pub fn list_auto_ml_jobs(&self) -> ListAutoMLJobs<C, M, R>

Constructs a fluent builder for the ListAutoMLJobs operation.

See ListAutoMLJobs for more information about the operation and its arguments.

pub fn list_candidates_for_auto_ml_job(
&self
) -> ListCandidatesForAutoMLJob<C, M, R>

Constructs a fluent builder for the ListCandidatesForAutoMLJob operation.

See ListCandidatesForAutoMLJob for more information about the operation and its arguments.

pub fn list_code_repositories(&self) -> ListCodeRepositories<C, M, R>

Constructs a fluent builder for the ListCodeRepositories operation.

See ListCodeRepositories for more information about the operation and its arguments.

pub fn list_compilation_jobs(&self) -> ListCompilationJobs<C, M, R>

Constructs a fluent builder for the ListCompilationJobs operation.

See ListCompilationJobs for more information about the operation and its arguments.

pub fn list_contexts(&self) -> ListContexts<C, M, R>

Constructs a fluent builder for the ListContexts operation.

See ListContexts for more information about the operation and its arguments.

pub fn list_data_quality_job_definitions(
&self
) -> ListDataQualityJobDefinitions<C, M, R>

Constructs a fluent builder for the ListDataQualityJobDefinitions operation.

See ListDataQualityJobDefinitions for more information about the operation and its arguments.

pub fn list_device_fleets(&self) -> ListDeviceFleets<C, M, R>

Constructs a fluent builder for the ListDeviceFleets operation.

See ListDeviceFleets for more information about the operation and its arguments.

pub fn list_devices(&self) -> ListDevices<C, M, R>

Constructs a fluent builder for the ListDevices operation.

See ListDevices for more information about the operation and its arguments.

pub fn list_domains(&self) -> ListDomains<C, M, R>

Constructs a fluent builder for the ListDomains operation.

See ListDomains for more information about the operation and its arguments.

pub fn list_edge_packaging_jobs(&self) -> ListEdgePackagingJobs<C, M, R>

Constructs a fluent builder for the ListEdgePackagingJobs operation.

See ListEdgePackagingJobs for more information about the operation and its arguments.

pub fn list_endpoint_configs(&self) -> ListEndpointConfigs<C, M, R>

Constructs a fluent builder for the ListEndpointConfigs operation.

See ListEndpointConfigs for more information about the operation and its arguments.

pub fn list_endpoints(&self) -> ListEndpoints<C, M, R>

Constructs a fluent builder for the ListEndpoints operation.

See ListEndpoints for more information about the operation and its arguments.

pub fn list_experiments(&self) -> ListExperiments<C, M, R>

Constructs a fluent builder for the ListExperiments operation.

See ListExperiments for more information about the operation and its arguments.

pub fn list_feature_groups(&self) -> ListFeatureGroups<C, M, R>

Constructs a fluent builder for the ListFeatureGroups operation.

See ListFeatureGroups for more information about the operation and its arguments.

pub fn list_flow_definitions(&self) -> ListFlowDefinitions<C, M, R>

Constructs a fluent builder for the ListFlowDefinitions operation.

See ListFlowDefinitions for more information about the operation and its arguments.

pub fn list_human_task_uis(&self) -> ListHumanTaskUis<C, M, R>

Constructs a fluent builder for the ListHumanTaskUis operation.

See ListHumanTaskUis for more information about the operation and its arguments.

pub fn list_hyper_parameter_tuning_jobs(
&self
) -> ListHyperParameterTuningJobs<C, M, R>

Constructs a fluent builder for the ListHyperParameterTuningJobs operation.

See ListHyperParameterTuningJobs for more information about the operation and its arguments.

pub fn list_images(&self) -> ListImages<C, M, R>

Constructs a fluent builder for the ListImages operation.

See ListImages for more information about the operation and its arguments.

pub fn list_image_versions(&self) -> ListImageVersions<C, M, R>

Constructs a fluent builder for the ListImageVersions operation.

See ListImageVersions for more information about the operation and its arguments.

pub fn list_labeling_jobs(&self) -> ListLabelingJobs<C, M, R>

Constructs a fluent builder for the ListLabelingJobs operation.

See ListLabelingJobs for more information about the operation and its arguments.

pub fn list_labeling_jobs_for_workteam(
&self
) -> ListLabelingJobsForWorkteam<C, M, R>

Constructs a fluent builder for the ListLabelingJobsForWorkteam operation.

See ListLabelingJobsForWorkteam for more information about the operation and its arguments.

pub fn list_model_bias_job_definitions(
&self
) -> ListModelBiasJobDefinitions<C, M, R>

Constructs a fluent builder for the ListModelBiasJobDefinitions operation.

See ListModelBiasJobDefinitions for more information about the operation and its arguments.

pub fn list_model_explainability_job_definitions(
&self
) -> ListModelExplainabilityJobDefinitions<C, M, R>

Constructs a fluent builder for the ListModelExplainabilityJobDefinitions operation.

See ListModelExplainabilityJobDefinitions for more information about the operation and its arguments.

pub fn list_model_package_groups(&self) -> ListModelPackageGroups<C, M, R>

Constructs a fluent builder for the ListModelPackageGroups operation.

See ListModelPackageGroups for more information about the operation and its arguments.

pub fn list_model_packages(&self) -> ListModelPackages<C, M, R>

Constructs a fluent builder for the ListModelPackages operation.

See ListModelPackages for more information about the operation and its arguments.

pub fn list_model_quality_job_definitions(
&self
) -> ListModelQualityJobDefinitions<C, M, R>

Constructs a fluent builder for the ListModelQualityJobDefinitions operation.

See ListModelQualityJobDefinitions for more information about the operation and its arguments.

pub fn list_models(&self) -> ListModels<C, M, R>

Constructs a fluent builder for the ListModels operation.

See ListModels for more information about the operation and its arguments.

pub fn list_monitoring_executions(&self) -> ListMonitoringExecutions<C, M, R>

Constructs a fluent builder for the ListMonitoringExecutions operation.

See ListMonitoringExecutions for more information about the operation and its arguments.

pub fn list_monitoring_schedules(&self) -> ListMonitoringSchedules<C, M, R>

Constructs a fluent builder for the ListMonitoringSchedules operation.

See ListMonitoringSchedules for more information about the operation and its arguments.

pub fn list_notebook_instance_lifecycle_configs(
&self
) -> ListNotebookInstanceLifecycleConfigs<C, M, R>

Constructs a fluent builder for the ListNotebookInstanceLifecycleConfigs operation.

See ListNotebookInstanceLifecycleConfigs for more information about the operation and its arguments.

pub fn list_notebook_instances(&self) -> ListNotebookInstances<C, M, R>

Constructs a fluent builder for the ListNotebookInstances operation.

See ListNotebookInstances for more information about the operation and its arguments.

pub fn list_pipeline_executions(&self) -> ListPipelineExecutions<C, M, R>

Constructs a fluent builder for the ListPipelineExecutions operation.

See ListPipelineExecutions for more information about the operation and its arguments.

pub fn list_pipeline_execution_steps(
&self
) -> ListPipelineExecutionSteps<C, M, R>

Constructs a fluent builder for the ListPipelineExecutionSteps operation.

See ListPipelineExecutionSteps for more information about the operation and its arguments.

pub fn list_pipeline_parameters_for_execution(
&self
) -> ListPipelineParametersForExecution<C, M, R>

Constructs a fluent builder for the ListPipelineParametersForExecution operation.

See ListPipelineParametersForExecution for more information about the operation and its arguments.

pub fn list_pipelines(&self) -> ListPipelines<C, M, R>

Constructs a fluent builder for the ListPipelines operation.

See ListPipelines for more information about the operation and its arguments.

pub fn list_processing_jobs(&self) -> ListProcessingJobs<C, M, R>

Constructs a fluent builder for the ListProcessingJobs operation.

See ListProcessingJobs for more information about the operation and its arguments.

pub fn list_projects(&self) -> ListProjects<C, M, R>

Constructs a fluent builder for the ListProjects operation.

See ListProjects for more information about the operation and its arguments.

pub fn list_studio_lifecycle_configs(
&self
) -> ListStudioLifecycleConfigs<C, M, R>

Constructs a fluent builder for the ListStudioLifecycleConfigs operation.

See ListStudioLifecycleConfigs for more information about the operation and its arguments.

pub fn list_subscribed_workteams(&self) -> ListSubscribedWorkteams<C, M, R>

Constructs a fluent builder for the ListSubscribedWorkteams operation.

See ListSubscribedWorkteams for more information about the operation and its arguments.

pub fn list_tags(&self) -> ListTags<C, M, R>

Constructs a fluent builder for the ListTags operation.

See ListTags for more information about the operation and its arguments.

pub fn list_training_jobs(&self) -> ListTrainingJobs<C, M, R>

Constructs a fluent builder for the ListTrainingJobs operation.

See ListTrainingJobs for more information about the operation and its arguments.

pub fn list_training_jobs_for_hyper_parameter_tuning_job(
&self
) -> ListTrainingJobsForHyperParameterTuningJob<C, M, R>

Constructs a fluent builder for the ListTrainingJobsForHyperParameterTuningJob operation.

See ListTrainingJobsForHyperParameterTuningJob for more information about the operation and its arguments.

pub fn list_transform_jobs(&self) -> ListTransformJobs<C, M, R>

Constructs a fluent builder for the ListTransformJobs operation.

See ListTransformJobs for more information about the operation and its arguments.

pub fn list_trial_components(&self) -> ListTrialComponents<C, M, R>

Constructs a fluent builder for the ListTrialComponents operation.

See ListTrialComponents for more information about the operation and its arguments.

pub fn list_trials(&self) -> ListTrials<C, M, R>

Constructs a fluent builder for the ListTrials operation.

See ListTrials for more information about the operation and its arguments.

pub fn list_user_profiles(&self) -> ListUserProfiles<C, M, R>

Constructs a fluent builder for the ListUserProfiles operation.

See ListUserProfiles for more information about the operation and its arguments.

pub fn list_workforces(&self) -> ListWorkforces<C, M, R>

Constructs a fluent builder for the ListWorkforces operation.

See ListWorkforces for more information about the operation and its arguments.

pub fn list_workteams(&self) -> ListWorkteams<C, M, R>

Constructs a fluent builder for the ListWorkteams operation.

See ListWorkteams for more information about the operation and its arguments.

pub fn put_model_package_group_policy(
&self
) -> PutModelPackageGroupPolicy<C, M, R>

Constructs a fluent builder for the PutModelPackageGroupPolicy operation.

See PutModelPackageGroupPolicy for more information about the operation and its arguments.

pub fn register_devices(&self) -> RegisterDevices<C, M, R>

Constructs a fluent builder for the RegisterDevices operation.

See RegisterDevices for more information about the operation and its arguments.

pub fn render_ui_template(&self) -> RenderUiTemplate<C, M, R>

Constructs a fluent builder for the RenderUiTemplate operation.

See RenderUiTemplate for more information about the operation and its arguments.

pub fn retry_pipeline_execution(&self) -> RetryPipelineExecution<C, M, R>

Constructs a fluent builder for the RetryPipelineExecution operation.

See RetryPipelineExecution for more information about the operation and its arguments.

pub fn search(&self) -> Search<C, M, R>

Constructs a fluent builder for the Search operation.

See Search for more information about the operation and its arguments.

pub fn send_pipeline_execution_step_failure(
&self
) -> SendPipelineExecutionStepFailure<C, M, R>

Constructs a fluent builder for the SendPipelineExecutionStepFailure operation.

See SendPipelineExecutionStepFailure for more information about the operation and its arguments.

pub fn send_pipeline_execution_step_success(
&self
) -> SendPipelineExecutionStepSuccess<C, M, R>

Constructs a fluent builder for the SendPipelineExecutionStepSuccess operation.

See SendPipelineExecutionStepSuccess for more information about the operation and its arguments.

pub fn start_monitoring_schedule(&self) -> StartMonitoringSchedule<C, M, R>

Constructs a fluent builder for the StartMonitoringSchedule operation.

See StartMonitoringSchedule for more information about the operation and its arguments.

pub fn start_notebook_instance(&self) -> StartNotebookInstance<C, M, R>

Constructs a fluent builder for the StartNotebookInstance operation.

See StartNotebookInstance for more information about the operation and its arguments.

pub fn start_pipeline_execution(&self) -> StartPipelineExecution<C, M, R>

Constructs a fluent builder for the StartPipelineExecution operation.

See StartPipelineExecution for more information about the operation and its arguments.

pub fn stop_auto_ml_job(&self) -> StopAutoMLJob<C, M, R>

Constructs a fluent builder for the StopAutoMLJob operation.

See StopAutoMLJob for more information about the operation and its arguments.

pub fn stop_compilation_job(&self) -> StopCompilationJob<C, M, R>

Constructs a fluent builder for the StopCompilationJob operation.

See StopCompilationJob for more information about the operation and its arguments.

pub fn stop_edge_packaging_job(&self) -> StopEdgePackagingJob<C, M, R>

Constructs a fluent builder for the StopEdgePackagingJob operation.

See StopEdgePackagingJob for more information about the operation and its arguments.

pub fn stop_hyper_parameter_tuning_job(
&self
) -> StopHyperParameterTuningJob<C, M, R>

Constructs a fluent builder for the StopHyperParameterTuningJob operation.

See StopHyperParameterTuningJob for more information about the operation and its arguments.

pub fn stop_labeling_job(&self) -> StopLabelingJob<C, M, R>

Constructs a fluent builder for the StopLabelingJob operation.

See StopLabelingJob for more information about the operation and its arguments.

pub fn stop_monitoring_schedule(&self) -> StopMonitoringSchedule<C, M, R>

Constructs a fluent builder for the StopMonitoringSchedule operation.

See StopMonitoringSchedule for more information about the operation and its arguments.

pub fn stop_notebook_instance(&self) -> StopNotebookInstance<C, M, R>

Constructs a fluent builder for the StopNotebookInstance operation.

See StopNotebookInstance for more information about the operation and its arguments.

pub fn stop_pipeline_execution(&self) -> StopPipelineExecution<C, M, R>

Constructs a fluent builder for the StopPipelineExecution operation.

See StopPipelineExecution for more information about the operation and its arguments.

pub fn stop_processing_job(&self) -> StopProcessingJob<C, M, R>

Constructs a fluent builder for the StopProcessingJob operation.

See StopProcessingJob for more information about the operation and its arguments.

pub fn stop_training_job(&self) -> StopTrainingJob<C, M, R>

Constructs a fluent builder for the StopTrainingJob operation.

See StopTrainingJob for more information about the operation and its arguments.

pub fn stop_transform_job(&self) -> StopTransformJob<C, M, R>

Constructs a fluent builder for the StopTransformJob operation.

See StopTransformJob for more information about the operation and its arguments.

pub fn update_action(&self) -> UpdateAction<C, M, R>

Constructs a fluent builder for the UpdateAction operation.

See UpdateAction for more information about the operation and its arguments.

pub fn update_app_image_config(&self) -> UpdateAppImageConfig<C, M, R>

Constructs a fluent builder for the UpdateAppImageConfig operation.

See UpdateAppImageConfig for more information about the operation and its arguments.

pub fn update_artifact(&self) -> UpdateArtifact<C, M, R>

Constructs a fluent builder for the UpdateArtifact operation.

See UpdateArtifact for more information about the operation and its arguments.

pub fn update_code_repository(&self) -> UpdateCodeRepository<C, M, R>

Constructs a fluent builder for the UpdateCodeRepository operation.

See UpdateCodeRepository for more information about the operation and its arguments.

pub fn update_context(&self) -> UpdateContext<C, M, R>

Constructs a fluent builder for the UpdateContext operation.

See UpdateContext for more information about the operation and its arguments.

pub fn update_device_fleet(&self) -> UpdateDeviceFleet<C, M, R>

Constructs a fluent builder for the UpdateDeviceFleet operation.

See UpdateDeviceFleet for more information about the operation and its arguments.

pub fn update_devices(&self) -> UpdateDevices<C, M, R>

Constructs a fluent builder for the UpdateDevices operation.

See UpdateDevices for more information about the operation and its arguments.

pub fn update_domain(&self) -> UpdateDomain<C, M, R>

Constructs a fluent builder for the UpdateDomain operation.

See UpdateDomain for more information about the operation and its arguments.

pub fn update_endpoint(&self) -> UpdateEndpoint<C, M, R>

Constructs a fluent builder for the UpdateEndpoint operation.

See UpdateEndpoint for more information about the operation and its arguments.

pub fn update_endpoint_weights_and_capacities(
&self
) -> UpdateEndpointWeightsAndCapacities<C, M, R>

Constructs a fluent builder for the UpdateEndpointWeightsAndCapacities operation.

See UpdateEndpointWeightsAndCapacities for more information about the operation and its arguments.

pub fn update_experiment(&self) -> UpdateExperiment<C, M, R>

Constructs a fluent builder for the UpdateExperiment operation.

See UpdateExperiment for more information about the operation and its arguments.

pub fn update_image(&self) -> UpdateImage<C, M, R>

Constructs a fluent builder for the UpdateImage operation.

See UpdateImage for more information about the operation and its arguments.

pub fn update_model_package(&self) -> UpdateModelPackage<C, M, R>

Constructs a fluent builder for the UpdateModelPackage operation.

See UpdateModelPackage for more information about the operation and its arguments.

pub fn update_monitoring_schedule(&self) -> UpdateMonitoringSchedule<C, M, R>

Constructs a fluent builder for the UpdateMonitoringSchedule operation.

See UpdateMonitoringSchedule for more information about the operation and its arguments.

pub fn update_notebook_instance(&self) -> UpdateNotebookInstance<C, M, R>

Constructs a fluent builder for the UpdateNotebookInstance operation.

See UpdateNotebookInstance for more information about the operation and its arguments.

pub fn update_notebook_instance_lifecycle_config(
&self
) -> UpdateNotebookInstanceLifecycleConfig<C, M, R>

Constructs a fluent builder for the UpdateNotebookInstanceLifecycleConfig operation.

See UpdateNotebookInstanceLifecycleConfig for more information about the operation and its arguments.

pub fn update_pipeline(&self) -> UpdatePipeline<C, M, R>

Constructs a fluent builder for the UpdatePipeline operation.

See UpdatePipeline for more information about the operation and its arguments.

pub fn update_pipeline_execution(&self) -> UpdatePipelineExecution<C, M, R>

Constructs a fluent builder for the UpdatePipelineExecution operation.

See UpdatePipelineExecution for more information about the operation and its arguments.

pub fn update_project(&self) -> UpdateProject<C, M, R>

Constructs a fluent builder for the UpdateProject operation.

See UpdateProject for more information about the operation and its arguments.

pub fn update_training_job(&self) -> UpdateTrainingJob<C, M, R>

Constructs a fluent builder for the UpdateTrainingJob operation.

See UpdateTrainingJob for more information about the operation and its arguments.

pub fn update_trial(&self) -> UpdateTrial<C, M, R>

Constructs a fluent builder for the UpdateTrial operation.

See UpdateTrial for more information about the operation and its arguments.

pub fn update_trial_component(&self) -> UpdateTrialComponent<C, M, R>

Constructs a fluent builder for the UpdateTrialComponent operation.

See UpdateTrialComponent for more information about the operation and its arguments.

pub fn update_user_profile(&self) -> UpdateUserProfile<C, M, R>

Constructs a fluent builder for the UpdateUserProfile operation.

See UpdateUserProfile for more information about the operation and its arguments.

pub fn update_workforce(&self) -> UpdateWorkforce<C, M, R>

Constructs a fluent builder for the UpdateWorkforce operation.

See UpdateWorkforce for more information about the operation and its arguments.

pub fn update_workteam(&self) -> UpdateWorkteam<C, M, R>

Constructs a fluent builder for the UpdateWorkteam operation.

See UpdateWorkteam for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(
conf: Config,
conn: C
) -> Client<C, AwsMiddleware, Standard>

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from the service Config.